requirements.txt
# pip install -r requirements.txt

matplotlib>=3.8.4
numpy>=1.26.4
opencv-python>=4.9.0.80
Pillow>=10.3.0
torch>=2.2.2
torchvision>=0.17.2
tqdm>=4.66.2
torchsummary>=1.5.1
